Move screen to component
Move screen to component
