+++ screens/linksScreen/css.js
@@ -1,0 +1,14 @@
+import {
+    StyleSheet,
+    Platform,
+  } from 'react-native';
+  
+  const styles = StyleSheet.create({
+    container: {
+      flex: 1,
+      paddingTop: 15,
+      backgroundColor: '#fff',
+    },
+  });
+
+  export default styles;
