+++ screens/settingsScreen/index.js
@@ -1,0 +1,18 @@
+import { connect } from 'react-redux';
+import SettingsScreen from './SettingsScreen';
+import * as  counterActions from '../../src/actions/counterActions'
+
+const mapStateToProps = state => {
+    return {
+        // counter: state.counterReducer.counter,
+    }
+  }
+
+const mapDispatchToProps = dispatch => {
+    return {
+        // increaseCounter: () => dispatch(counterActions.increaseCounter()),
+        // decreaseCounter: () => dispatch(counterActions.decreaseCounter()),
+    }
+  }
+  
+  export default connect(mapStateToProps, mapDispatchToProps)(SettingsScreen);
